refactor: [step2] Scores 추상화 완료
